requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
great_expectations
pyyaml
